use venv

+++ installer/build-windows.ps1
@@ -18,8 +18,11 @@ unzip ZorinBlue.zip
 curl https://skytemple.org/build_deps/nsis.zip -O
 unzip -o nsis.zip -d "C:\Program Files (x86)\NSIS"
 
+python -m venv C:\skytemple-venv
+C:\skytemple-venv\Scripts\activate.ps1
+
 # Install PyInstaller
-pip3 install setuptools wheel pyinstaller
+pip install setuptools wheel pyinstaller
 
 # Install PyGObject and pycairo
 pip install --force-reinstall (Resolve-Path C:\gtk-build\build\x64\release\pygobject\dist\PyGObject*.whl)
@@ -38,11 +41,15 @@ if (-not $env:IS_DEV_BUILD) {
 
 pyinstaller skytemple.spec
 
+if(!(Test-Path ".\dist\skytemple\skytemple.exe")){
+    return 1
+}
+
 # Check if we need to copy the cacert file
 if (Test-Path ".\dist\skytemple\certifi\cacert.pem") {
   echo "Moved cacert to correct place"
   cp -rf dist/skytemple/certifi/cacert.pem dist/skytemple/certifi.pem
 }
 
-echo $env:PACKAGE_VERSION | Out-File -FilePath $env:GITHUB_ENV -Encoding utf8
-echo $env:PACKAGE_VERSION | Out-File -FilePath $env:GITHUB_ENV -Encoding utf8
+echo $env:PACKAGE_VERSION | Out-File -FilePath dist/skytemple/VERSION -Encoding utf8
+echo $env:PACKAGE_VERSION | Out-File -FilePath dist/skytemple/data/VERSION -Encoding utf8
